Add X_FORWARDED_PROTO to LiveReload#protocol
This commit adds `@env['HTTP_HTTP_X_FORWARDED_PROTO']`
to the list of items that `LiveReload#protocol` can
pull from to populate http/https from.

This is useful in situations where you have a dev machine
you are doing work on, but are also reverse-proxying back
to that machine via HTTPS to give access to another device,
that you can't control mixed content on, but still need to
be able to access via your dev box locally.

Signed-off-by: Evan Gray <evanscottgray@gmail.com>